Link to SMS convos from all user detail pages.
